// src/ReduxTool/Slice/Consumer/types.ts
export interface IOrderProjectType {
    address: string,
    area: number
    consumerCategory: null | string
    createddt: string
    dc: 0
    graphicLayer: string
    installationmode: "PVNxt Mode"|"Self Mode"|""
    iscompleted?: boolean
    lat: number
    lng: number
    loading: number
    userid: string
    plantcapacity: number
    projectid: string
    projectname: string
    projecttype: string
    reduceElectricityBill: boolean
    totalroofarea: number
    updateddt: null | string
    user: null | string
    sanctionload:number
}